requirements.txt
python-telegram-bot==21.6
Pillow==10.4.0